заменил исчесление скорости с int на double так более гибко настраивать скорость можно

+++ src/main/java/org/example/deadknight/components/SpeedComponent.java
@@ -7,7 +7,7 @@ public class SpeedComponent extends Component implements HasSpeed {
 
     private double speed;
 
-    public SpeedComponent(int speed) {
+    public SpeedComponent(double speed) {
         this.speed = speed;
     }
